+++ lib/opam_solve.mli
@@ -13,7 +13,8 @@ val local_opam_config_solver : (switch, switch_diagnostics) t
 val explicit_repos_solver :
   (opam_env * explicit_repos, explicit_repos_diagnostics) t
 
-val mock_solver : (opam_env * OpamFile.OPAM.t list, mock_diagnostics) t
+val mock_solver :
+  (opam_env * OpamFile.OPAM.t list * OpamPackage.t list, mock_diagnostics) t
 
 val calculate :
   build_only:bool ->